requirements.txt
numpy==2.0.1
pillow==10.4.0
opencv-python==4.10.0.84
imutils==0.5.4